+++ Libraries/src/Graphics/Texture.cpp
@@ -103,13 +103,13 @@ void Texture::SetTextureData(void *data, int width, int height, GLenum format, G
     glGenTextures(1, &this->ID);
     this->Bind();
 
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, filter);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, filter);
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, static_cast<GLint>(filter));
+    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, static_cast<GLint>(filter));
 
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, this->format, this->Width, this->Height, 0, this->format, pixelType, data);
+    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(this->format), this->Width, this->Height, 0, this->format, pixelType, data);
     glGenerateMipmap(GL_TEXTURE_2D);
 
     this->Unbind();
@@ -189,7 +189,7 @@ void Texture::SetFramebufferTexture(const char *uniformName, GLuint slot, int wi
 
     glGenTextures(1, &this->ID);
     glBindTexture(GL_TEXTURE_2D, this->ID);
-    glTexImage2D(GL_TEXTURE_2D, 0, this->format, this->Width, this->Height, 0, this->format, this->pixelType, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(this->format), this->Width, this->Height, 0, this->format, this->pixelType, nullptr);
 
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
@@ -208,14 +208,14 @@ void Texture::ResizeFramebufferTexture(int width, int height)
     this->Width = width;
     this->Height = height;
     glBindTexture(GL_TEXTURE_2D, this->ID);
-    glTexImage2D(GL_TEXTURE_2D, 0, this->format, this->Width, this->Height, 0, this->format, GL_UNSIGNED_BYTE, nullptr);
+    glTexImage2D(GL_TEXTURE_2D, 0, static_cast<GLint>(this->format), this->Width, this->Height, 0, this->format, GL_UNSIGNED_BYTE, nullptr);
     glBindTexture(GL_TEXTURE_2D, 0);
 }
 
 void Texture::texUnit(const Shader &shader) const
 {
     shader.Bind();
-    glUniform1i(glGetUniformLocation(shader.GetID(), this->UniformName), this->slot);
+    glUniform1i(glGetUniformLocation(shader.GetID(), this->UniformName), static_cast<GLint>(this->slot));
 }
 
 void Texture::Bind() const
